refactor: remove unused UserMenu, clean up WaterColumn, migrate AdoptTree to components/trees, rename SearchTrees to NearbyTrees
