+++ frontend/src/components/dev/NearbyTrees.svelte
