+++ frontend/src/components/search/FilterTrees.svelte
