+++ frontend/src/components/status/AdoptedTrees.svelte
