+++ frontend/src/components/status/NearbyTrees.svelte
