+++ frontend/src/components/status/index.ts
@@ -1,0 +1,2 @@
+export { default as AdoptedTrees } from './AdoptedTrees.svelte';
+export { default as NearbyTrees } from './NearbyTrees.svelte';
