+++ frontend/src/components/trees/AdoptTreeButton.svelte
