+++ frontend/src/routes/trees/[treeId]/+page.svelte
@@ -4,7 +4,7 @@
 	import Chat from '$components/chat/Chat.svelte';
 	import { DialogPanel } from '$components/overlay';
 
-	import AdoptTree from '$components/trees/AdoptTree.svelte';
+	import AdoptTree from '$components/trees/AdoptTreeButton.svelte';
 	import { supabase } from '../../../supabase';
 	import { onMount } from 'svelte';
 	import { page } from '$app/stores';
